requirements.txt
absl-py==2.0.0
aiosignal==1.3.1
altair==5.4.1
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
backports.functools_lru_cache==1.6.5
biopython==1.81
blinker==1.8.2
bokeh==3.3.0
cachetools==5.3.2
catboost==1.2.5
certifi==2023.7.22
charset_normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
et_xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google_auth==2.23.4
google_auth_oauthlib==1.0.0
google_pasta==0.2.0
graphviz==0.20.3
grpcio==1.59.2
h5py==3.10.0
idna==3.4
imageio==2.34.2
importlib_metadata==6.8.0
importlib_resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema_specifications==2023.12.1
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.9
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.4.0
llvmlite==0.43.0
LogUniform==2.0.1
Markdown==3.5.1
markdown_it_py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib_inline==0.1.6
mdurl==0.1.2
mglearn==0.2.0
ml_dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
narwhals==1.8.3
nest_asyncio==1.5.8
networkx==3.2.1
numba==0.60.0
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.2
opt_einsum==3.3.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3.1
platformdirs==3.11.0
plotly==5.22.0
prompt_toolkit==3.0.39
propythia==3.0.2
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.0
pyasn1_modules==0.3.0
pydeck==0.9.1
pydot==3.0.1
pyfaidx==0.7.2.2
pygame==2.1.3
Pygments==2.16.1
pyparsing==3.1.1
python_box==7.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
ray==2.9.0
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.8.1
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==68.2.2
shap==0.45.1
six==1.16.0
slicer==0.0.8
smart-open==6.4.0
smmap==5.0.1
squiggle==0.3.1
stack-data==0.6.2
streamlit==1.38.0
sympy==1.12
tenacity==8.4.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wheel==0.41.3
widgetsnbextension==4.0.9
wrapt==1.14.1
xgboost==2.1.0
xyzservices==2023.10.1
zipp==3.17.0
